+++ test/gen_time_zone_database/zone_info_reader_test.dart
@@ -78,7 +78,8 @@ void main() {
           laTransitions,
           contains(ZoneTransition((b) => b
             ..transitionTime =
-                OffsetDateTime(ZoneOffset(-8), 2000, 4, 2, 2).toInstant()
+                OffsetDateTime.withOffset(ZoneOffset(-8), 2000, 4, 2, 2)
+                    .toInstant()
             ..offset = NamedZoneOffset('PDT', true, -7))));
     });